ConfigLocators to extend common parent
XmlConfigLocator and XmlClientConfigLocator to extend
AbstractConfigLocator that can be used as a basis for YAML counterparts
as well.
